+++ eliot/__init__.py
@@ -9,6 +9,11 @@ from ._action import (
     start_action, startTask, Action, preserve_context, current_action,
     use_asyncio_context, log_call
 )
+from ._generators import (
+    GeneratorSupportNotEnabled,
+    use_generator_context,
+    eliot_friendly_generator_function,
+)
 from ._output import (
     ILogger,
     Logger,
@@ -83,6 +88,10 @@ __all__ = [
     "current_action",
     "use_asyncio_context",
     "ValidationError",
+
+    "GeneratorSupportNotEnabled",
+    "use_generator_context",
+    "eliot_friendly_generator_function",
 
     # PEP 8 variants:
     "write_traceback",